backend/requirements.txt
fastapi
uvicorn
transformers
torch
python-dotenv==1.0.1
requests==2.31.0